Move KernelInfoState and AAKernelInfo to llvm/include/llvm/Transforms/IPO/OpenMPOpt.h
Move functions to AAKernelInfo

Fix formatting issues

Update OpenMPOpt.cpp

Update OpenMPOpt.cpp
